+++ lib/boleto_holmes/log.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative('../utils/resource')
+require('starkcore')
 require_relative('../utils/rest')
-require_relative('../utils/checks')
 require_relative('boleto_holmes')
+
 
 module StarkBank
   class BoletoHolmes
@@ -19,13 +19,13 @@ module StarkBank
     # - holmes [BoletoHolmes]: BoletoHolmes entity to which the log refers to.
     # - type [string]: type of the Boleto event which triggered the log creation. ex: 'registered' or 'paid'
     # - created [DateTime]: creation datetime for the log. ex: DateTime.new(2020, 3, 10, 10, 30, 0, 0)
-    class Log < StarkBank::Utils::Resource
+    class Log < StarkCore::Utils::Resource
       attr_reader :id, :holmes, :type, :created
       def initialize(id:, holmes:, type:, created:)
         super(id)
         @holmes = holmes
         @type = type
-        @created = StarkBank::Utils::Checks.check_datetime(created)
+        @created = StarkCore::Utils::Checks.check_datetime(created)
       end
 
       # # Retrieve a specific Log
@@ -59,8 +59,8 @@ module StarkBank
       # ## Return:
       # - list of Log objects with updated attributes
       def self.query(limit: nil, after: nil, before: nil, types: nil, holmes_ids: nil, user: nil)
-        after = StarkBank::Utils::Checks.check_date(after)
-        before = StarkBank::Utils::Checks.check_date(before)
+        after = StarkCore::Utils::Checks.check_date(after)
+        before = StarkCore::Utils::Checks.check_date(before)
         StarkBank::Utils::Rest.get_stream(
           limit: limit,
           after: after,
@@ -89,8 +89,8 @@ module StarkBank
       # ## Return:
       # - list of Log objects with updated attributes and cursor to retrieve the next page of Log objects
       def self.page(cursor: nil, limit: nil, after: nil, before: nil, types: nil, holmes_ids: nil, user: nil)
-        after = StarkBank::Utils::Checks.check_date(after)
-        before = StarkBank::Utils::Checks.check_date(before)
+        after = StarkCore::Utils::Checks.check_date(after)
+        before = StarkCore::Utils::Checks.check_date(before)
         return StarkBank::Utils::Rest.get_page(
           cursor: cursor,
           limit: limit,
@@ -110,7 +110,7 @@ module StarkBank
           resource_maker: proc { |json|
             Log.new(
               id: json['id'],
-              holmes: StarkBank::Utils::API.from_api_json(holmes_maker, json['holmes']),
+              holmes: StarkCore::Utils::API.from_api_json(holmes_maker, json['holmes']),
               type: json['type'],
               created: json['created']
             )