+++ lib/workspace/workspace.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative('../utils/resource')
+require('starkcore')
 require_relative('../utils/rest')
-require_relative('../utils/checks')
+
 
 module StarkBank
   # # Workspace object
@@ -20,7 +20,7 @@ module StarkBank
   #
   # ## Attributes:
   # - id [string, default nil]: unique id returned when the workspace is created. ex: '5656565656565656'
-  class Workspace < StarkBank::Utils::Resource
+  class Workspace < StarkCore::Utils::Resource
     attr_reader :username, :name, :allowed_tax_ids, :id
     def initialize(username:, name:, allowed_tax_ids: nil, id: nil)
       super(id)